refactor: reorganize code in the gameboard
- Moved the `GameTile` class to its own module in `gametile.py`.
- Removed the `NEIGHBOR_OFFSETS` constant from the `gameboard.py` file.
- Cleaned up and reformatted the code for better readability.

This commit improves code organization and maintainability.
