+++ src/modules/gameboard/path.py
@@ -1,5 +1,6 @@
 from typing import List, Tuple
-from src.modules.gameboard.gameboard import GameTile
+
+from src.modules.gameboard.gametile import GameTile
 
 
 class Path:
