+++ src/modules/gameboard/resultlist.py
@@ -1,5 +1,6 @@
 from typing import Any, Dict, Generator, List, Tuple
-from src.modules.gameboard.gameboard import GameTile
+
+from src.modules.gameboard.gametile import GameTile
 from src.utils.timer import Timer
